week6/requirements.txt
dotenv==0.9.9
fastapi==0.132.0
fastjsonschema==2.21.2
httpcore==1.0.9
httpx==0.28.1
imbalanced-learn==0.14.1
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
matplotlib==3.10.8
matplotlib-inline==0.2.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.3
notebook_shim==0.2.4
numba==0.64.0
numpy==2.4.2
optuna==4.7.0
pandas==3.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
requests==2.32.5
scikit-learn==1.8.0
seaborn==0.13.2
setuptools==82.0.0
shap==0.50.0
sklearn-compat==0.1.5
uvicorn==0.41.0
xgboost==3.2.0
